Bug fix for no ecu connected
When the ecu was not connected the websocket was brain farting, seems that we are creating kpro and s300 instances
too fast trying to reconnect and for some reason the websocket was not really working or not sending anything to the frontend.
Commit edaca902623b5dc75bb49fe137cc42a73ec992e1 was supposed to be version 3.0.0 but this new one would overwrite it since
the other was has the bug and I don´t really want to make a new release.